0.2.0 upload
5 Levels
Coins
Loot boxes are currently in the game and you may open them but they are just a button that you push. It does not give you a reward. It does take your coins though.


In the next update I plan on finishing up the loot crates and inventory then start modeling some loot.
After that I plan to add an infinite run mode.